chips: add shared qemu_virt_chip
