+++ chips/qemu_virt_chip/README.md
@@ -1,0 +1,5 @@
+qemu-system-riscv* `virt` machine chip crate
+=============================================
+
+Shared peripherals for QEMU RISC-V `virt` chips.
+
