+++ chips/qemu_virt_chip/src/lib.rs
@@ -1,0 +1,10 @@
+// Licensed under the Apache License, Version 2.0 or the MIT License.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Copyright Tock Contributors 2026.
+
+//! Peripherals for the qemu-system-riscv* virt machines
+
+#![forbid(unsafe_code)]
+#![no_std]
+
+pub mod uart;
